Refact: Cambio de nombre del archivo y refactor de la clase
